requirements.txt
Django==5.2  # Or your specific Django version
djangorestframework
psycopg2-binary  # PostgreSQL connector
corsheaders
Pillow # For image handling
whitenoise # For serving static files
django-environ # For easy environment variable management (recommended)
# Add any other packages you need ...